requirements.txt
# Core ML Libraries
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.2

# Data Processing
scipy==1.11.1
imbalanced-learn==0.11.0
feature-engine==1.6.2

# Explainable AI
shap==0.42.1
lime==0.2.0.1
eli5==0.13.0

# API Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6

# Dashboard
streamlit==1.26.0
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2

# MLOps
mlflow==2.6.0
optuna==3.3.0
hydra-core==1.3.2

# Database
pymongo==4.5.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7

# Cloud & Storage
boto3==1.28.47
azure-storage-blob==12.17.0
google-cloud-storage==2.10.0

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
tqdm==4.66.1
joblib==1.3.2

# Monitoring & Logging
prometheus-client==0.17.1
structlog==23.1.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.24.1

# Security
cryptography==41.0.4

# other
xlrd
ensure
python-box